[bios] Use size_t when casting _text16_memsz and _data16_memsz

+++ src/arch/i386/firmware/pcbios/hidemem.c
@@ -76,9 +76,9 @@ extern struct segoff __text16 ( int15_vector );
 extern char _textdata[];
 extern char _etextdata[];
 extern char _text16_memsz[];
-#define _text16_memsz ( ( unsigned int ) _text16_memsz )
+#define _text16_memsz ( ( size_t ) _text16_memsz )
 extern char _data16_memsz[];
-#define _data16_memsz ( ( unsigned int ) _data16_memsz )
+#define _data16_memsz ( ( size_t ) _data16_memsz )
 
 /**
  * Hide region of memory from system memory map
